Add a few commented-out lines to terminate $ in code
A syntax-highlighting editor sees the $ as the start of a formula.
Within a `code` block, however, it actually doesn't get interpreted as
the start of a formula. As a result, the editor will highlight
everything to the end of a file as a formula.

Terminate the $ by adding another one. Putting it in a normal comment
doesn't trick the syntax highligher, so use \iffalse as a
pseudo-comment.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>