/module-alias-usage.dia: fix layout
When I open this figure, the 'the asix module is loaded' star appears on
the very top right, far away from the figure itself. As a result, that
star is the only thing shown on the slide.

Move the star down to the position I _think_ is the proper one.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>